% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACD.r
\name{ACD}
\alias{ACD}
\title{Accumulate Collection & Distribution}
\usage{
ACD(HLC, ...)
}
\arguments{
\item{HLC}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
Accumulate Collection & Distribution
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

