% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmf.r
\name{Cmf}
\alias{Cmf}
\title{Chaikin Money Flow}
\usage{
Cmf(OHLCV, n = 20)
}
\arguments{
\item{OHLCV}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n}{Number of periods for moving average.}
}
\description{
Chaikin Money Flow
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

