% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmo.r
\name{Cmo}
\alias{Cmo}
\title{Chande Momentum Oscillator (different from TTR)}
\usage{
Cmo(x, n = 14, m = 1)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n}{Number of periods for moving average.}

\item{m}{A function or a string naming the function to be called.}
}
\description{
Chande Momentum Oscillator (different from TTR)
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

