% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DI.r
\name{DI}
\alias{DI}
\title{Divergence Index}
\usage{
DI(price, fast = 5, slow = 20, n = 10, f = 2, m = 1, maType, ...)
}
\arguments{
\item{price}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{fast}{...}

\item{slow}{...}

\item{n}{Number of periods for moving average.}

\item{f}{...}

\item{m}{...}

\item{maType}{A function or a string naming the function to be called.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
Divergence Index
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

