% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DYMI.r
\name{DYMI}
\alias{DYMI}
\title{Dynamic Relative Momentum Index}
\usage{
DYMI(price, n = 14, d = 5, m = 1, maType, ...)
}
\arguments{
\item{price}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n}{Number of periods for moving average.}

\item{d}{...}

\item{m}{...}

\item{maType}{A function or a string naming the function to be called.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
Dynamic Relative Momentum Index
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

