% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KST.r
\name{KST}
\alias{KST}
\title{Know Sure Thing}
\usage{
KST(x, n1 = 10, n2 = 13, n3 = 15, n4 = 20, maType, ...)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n1}{Number of periods for moving average.}

\item{n2}{...}

\item{n3}{...}

\item{n4}{...}

\item{maType}{A function or a string naming the function to be called.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
Know Sure Thing
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

