% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA.r
\name{MAMA}
\alias{MAMA}
\title{MESA Adaptive Moving Average}
\usage{
MAMA(x, fastlimit = 0.5, slowlimit = 0.05)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{fastlimit}{...}

\item{slowlimit}{...}
}
\description{
MESA Adaptive Moving Average
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

