% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBands.r
\name{MBBands}
\alias{MBBands}
\title{Modified Bollinger Bands}
\usage{
MBBands(x, n = 20, f = 2)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n}{Number of periods for moving average.}

\item{f}{...}
}
\description{
Modified Bollinger Bands
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

