% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momentum.r
\name{Momentum}
\alias{Momentum}
\title{Momentum (more flexible)}
\usage{
Momentum(p1, p2 = p1, nlag, type = "diff", lable = NULL, maType = NULL,
  ...)
}
\arguments{
\item{p1}{first price column}

\item{p2}{second price column}

\item{nlag}{lag period}

\item{type}{momentum type "diff" or "ratio"}

\item{lable}{name of the momentum}

\item{maType}{A function or a string naming the function to be called.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
Momentum (more flexible)
}
\author{
Chen Chaozong
}

