% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momentum.r
\name{MomRatio}
\alias{MomRatio}
\title{MomRatio (more flexible momentum ratio)}
\usage{
MomRatio(positive, negative = positive, type = "positive", maType,
  lable = NULL, ratiomaType = NULL, ...)
}
\arguments{
\item{positive}{first price column}

\item{negative}{second price column}

\item{type}{momentum type in "positive","negative","net"}

\item{maType}{A function or a string naming the function to be called. for positive and negative momentum}

\item{lable}{name of the momentum}

\item{ratiomaType}{A function or a string naming the function to be called. for numerator and denominator}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
MomRatio (more flexible momentum ratio)
}
\author{
Chen Chaozong
}

