% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI.r
\name{SI}
\alias{SI}
\title{Stochastic Index}
\usage{
SI(HLC, nFastK = 14, nFastD = 3, nSlowD = 3, maType, bounded = TRUE,
  smooth = 1, ...)
}
\arguments{
\item{HLC}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{nFastK}{Number of periods for moving average.}

\item{nFastD}{...}

\item{nSlowD}{...}

\item{maType}{A function or a string naming the function to be called.}

\item{bounded}{...}

\item{smooth}{...}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
Stochastic Index
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

