% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMA.r
\name{TEMA}
\alias{TEMA}
\title{Triple exponential moving average}
\usage{
TEMA(x, n = 20, wilder = FALSE, ratio = NULL)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n}{Number of periods for moving average.}

\item{wilder}{...}

\item{ratio}{...}
}
\description{
Triple exponential moving average
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

