% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSI.r
\name{TSI}
\alias{TSI}
\title{True Strength Index}
\usage{
TSI(x, n = 20, m = 6, d = 1, maType, ...)
}
\arguments{
\item{x}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.}

\item{n}{Number of periods for moving average.}

\item{m}{...}

\item{d}{...}

\item{maType}{A function or a string naming the function to be called.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\description{
True Strength Index
}
\author{
Chen Chaozong
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.  See \code{\link{ATR}}, which uses true
range.  See \code{\link{chaikinVolatility}} for another volatility measure.
}
\keyword{ts}

